update ui,user bisa edit profile, add list-lelang & dashboard masyarakat,update tampilan penawaran, add dompdf untuk download laporan,add animate.css,update penggunaan count() skrg menggunakan eloquent, add password reveal di login & register, mengganti tipe data field harga dengan bigInt agar tidak terlimit

+++ database/migrations/2023_02_05_055526_create_barangs_table.php
@@ -17,7 +17,7 @@ class CreateBarangsTable extends Migration
             $table->id();
             $table->string('nama_barang');
             $table->date('tanggal');
-            $table->integer('harga_awal');
+            $table->bigInteger('harga_awal');
             $table->string('image')->nullable();
             $table->string('deskripsi_barang');
             $table->timestamps();
